mvp of extracting functions and bundling ts file
