+++ __tests__/pathToDotNotation.test.js
@@ -1,4 +1,4 @@
-const pathToDotNotation = require('../build.js').pathToDotNotation
+const pathToDotNotation = require('../dist/build.js').pathToDotNotation
 
 describe('Token formatters: dot.notation', () => {
   it('returns a valid token name in dot.notation', () => {
