+++ newToken.tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "compilerOptions": {
+    "outDir": "dist",
+    "target": "es2019",
+    "lib": [
+      "es2019"
+    ],
+    "module": "commonjs",
+    "esModuleInterop": true,
+    "strict": true,
+    "declaration": false,
+    "sourceMap": false,
+    "include": [
+      "build.ts"
+    ]
+  }
+}
