+++ src/@types/token.d.ts
@@ -1,0 +1,3 @@
+export type Token = {
+  path: string[]
+}
